refactor: move shapes into game
