+++ frontend/src/game/character/states.ts
@@ -1,5 +1,5 @@
 import { StateConfig } from '../entity';
-import { Rectangle } from '../../shapes';
+import { Rectangle } from '../shapes';
 import Box from '../box';
 
 import walk from '../sprites/character/walk';
