+++ frontend/src/game/entity/Placeholder/index.ts
@@ -1,5 +1,5 @@
 import { Entity } from '..';
-import { Point } from '../../../shapes';
+import { Point } from '../../shapes';
 import { PlaceholderState, states } from './states';
 
 class Placeholder extends Entity {
