+++ frontend/src/game/entity/types/entityClass.ts
@@ -1,4 +1,4 @@
-import { Point } from '../../../shapes';
+import { Point } from '../../shapes';
 
 type EntityClass<A> = new(position:Point)=>A;
