+++ frontend/src/game/levels/index.ts
@@ -1,7 +1,7 @@
 import { LevelLoad as Load, LevelId } from './types';
 import { LevelConfig, LoadingConfig as LoadZone, EntityConfig } from './levelConfig';
 
-import { Point, Line, Rectangle } from '../../shapes';
+import { Point, Line, Rectangle } from '../shapes';
 
 import { Entity, EntityClass, Direction } from '../entity';
 import entitiesList from '../entity/list';
