+++ frontend/src/game/levels/testLevel/index.ts
@@ -1,4 +1,4 @@
-import { Point, Rectangle } from '../../../shapes';
+import { Point, Rectangle } from '../../shapes';
 
 import {
   EntityConfig, LevelConfig, LoadingConfig, SurfaceConfig, surfaceConfigList,
