+++ frontend/src/game/shapes/point.ts
