+++ frontend/src/game/shapes/rectangle.ts
